Refactor project structure
- Move entry point to root
- Move setup to root
- Modularize other components
