+++ mtgapi/db_manager.py
